chore: deprecate old client config
